requirements.txt
pyserial
opencv-python
wmi; platform_system=="Windows"